Extract 'printLocation' & 'printSourceLocation' functions
Context: https://github.com/facebook/relay/pull/2752#issuecomment-498667289
