+++ src/error/index.js
@@ -1,12 +1,10 @@
 // @flow strict
 
-export { GraphQLError } from './GraphQLError';
+export { GraphQLError, printError } from './GraphQLError';
 
 export { syntaxError } from './syntaxError';
 
 export { locatedError } from './locatedError';
-
-export { printError } from './printError';
 
 export { formatError } from './formatError';
 export type { GraphQLFormattedError } from './formatError';
